# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de EPIs.",
        "title": "SaaS EPI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.radaptech.com.br",
            "email": "suporte@seusaas.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cadastro-departamento": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra um novo departamento no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Criar novo departamento",
                "parameters": [
                    {
                        "description": "Dados do departamento",
                        "name": "departamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Departamento"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Departamento criado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Departamento já existe",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/departamento/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um departamento pelo ID",
                "tags": [
                    "Departamentos"
                ],
                "summary": "Deletar departamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Departamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem Conteúdo"
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/departamentos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista paginada de departamentos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Listar departamentos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Página da lista",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade por página",
                        "name": "quantidade",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.DepartamentoPaginado"
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/departamentos/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Faz upload de uma planilha .xlsx para cadastrar múltiplos departamentos",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Importar departamentos via XLSX",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo de planilha (.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso na importação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Arquivo inválido ou vazio",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Sessão inválida",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/departamentos/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza o nome de um departamento existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Atualizar departamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Departamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do departamento",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Departamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Departamento": {
            "type": "object",
            "required": [
                "departamento"
            ],
            "properties": {
                "departamento": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "model.DepartamentoDto": {
            "type": "object",
            "properties": {
                "departamento": {
                    "type": "string",
                    "example": "Recursos Humanos"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "service.DepartamentoPaginado": {
            "type": "object",
            "properties": {
                "departamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DepartamentoDto"
                    }
                },
                "pagina": {
                    "type": "integer"
                },
                "paginaFinal": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}